+++ concordium-sdk/src/main/java/com/concordium/sdk/transactions/TransactionFactory.java
@@ -35,4 +35,36 @@ public class TransactionFactory {
     public static RegisterDataTransaction.RegisterDataTransactionBuilder newRegisterData() {
         return RegisterDataTransaction.builder();
     }
+
+    /**
+     * Creates a new {@link InitContractTransaction.InitContractTransactionBuilder} for
+     * creating a {@link InitContractTransaction}
+     *
+     * @return the builder for a {@link InitContractTransaction}
+     */
+    public static InitContractTransaction.InitContractTransactionBuilder newInitContract() {
+        return InitContractTransaction.builder();
+    }
+
+
+    /**
+     * Creates a new {@link UpdateContractTransaction.UpdateContractTransactionBuilder} for
+     * creating a {@link UpdateContractTransaction}
+     *
+     * @return the builder for a {@link UpdateContractTransaction}
+     */
+    public static UpdateContractTransaction.UpdateContractTransactionBuilder newUpdateContract() {
+        return UpdateContractTransaction.builder();
+    }
+
+    /**
+     * Creates a new {@link DeployModuleTransaction.DeployModuleTransactionBuilder} for
+     * creating a {@link DeployModuleTransaction}
+     *
+     * @return the builder for a {@link DeployModuleTransaction}
+     */
+    public static DeployModuleTransaction.DeployModuleTransactionBuilder newDeployModule() {
+        return DeployModuleTransaction.builder();
+    }
+
 }